requirements.txt
cryptg
Telethon
python-dotenv
PyPDF2

